fix: styling

+++ crates/disassemble/src/core/mod.rs
@@ -2,7 +2,7 @@ use std::time::Instant;
 
 use crate::{error::Error, interfaces::DisassemblerArgs};
 use eyre::eyre;
-use heimdall_common::utils::strings::{encode_hex, decode_hex};
+use heimdall_common::utils::strings::{decode_hex, encode_hex};
 use heimdall_vm::core::opcodes::opcode_name;
 use tracing::{debug, info};
 
@@ -28,16 +28,18 @@ pub async fn disassemble(args: DisassemblerArgs) -> Result<String, Error> {
     // get the bytecode from the target
     let start_fetch_time = Instant::now();
     debug!("fetching target bytecode took {:?}", start_fetch_time.elapsed());
-    
+
     // avoid the special case when the target length is exactly 20 bytes
     let contract_bytecode;
     if args.rpc_url == String::new() {
         contract_bytecode = decode_hex(&args.target).unwrap();
-    }else {
-        contract_bytecode =
-        args.get_bytecode().await.map_err(|e| eyre!("fetching target bytecode failed: {}", e))?;
+    } else {
+        contract_bytecode = args
+            .get_bytecode()
+            .await
+            .map_err(|e| eyre!("fetching target bytecode failed: {}", e))?;
     }
-    
+
     // iterate over the bytecode, disassembling each instruction
     let start_disassemble_time = Instant::now();
     while program_counter < contract_bytecode.len() {
@@ -68,7 +70,7 @@ pub async fn disassemble(args: DisassemblerArgs) -> Result<String, Error> {
             )
             .as_str(),
         );
-        program_counter += 1 + byte_count_to_push_offset as usize;
+        program_counter += 1 + byte_count_to_push_offset;
     }
     debug!("disassembly took {:?}", start_disassemble_time.elapsed());
